Working statistics module

+++ chemclipse/plugins/org.eclipse.chemclipse.numeric/src/org/eclipse/chemclipse/numeric/statistics/model/AnovaStatistics.java
@@ -36,7 +36,7 @@ public class AnovaStatistics implements IAnovaStatistics {
 					for(int i = 0; i < size; i++) {
 						T t = statisticsElements.get(i);
 						try {
-							values[i] = (double)getdata.invoke(t, new Object[0]);
+							values[i] = (float)getdata.invoke(t, new Object[0]);
 						} catch(IllegalAccessException e) {
 							e.printStackTrace();
 						} catch(IllegalArgumentException e) {
